feat(user): restructure user entity imports and add user entity definition feat(session): update session membership and room entities to reflect new user entity path feat(vscode): add launch configuration for debugging fix(package): update package-lock.json to use official npm registry fix(tsconfig): add path mapping for src directory
