+++ server/src/user/entities/user.entity.ts
