+++ server/src/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '../entities/user.entities';
+import { User } from './entities/user.entity';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
